+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_extracting_callbacks/prefer_extracting_callbacks_rule.dart
@@ -6,6 +6,7 @@ import 'package:analyzer/dart/ast/visitor.dart';
 import '../../../../../utils/flutter_types_utils.dart';
 import '../../../../../utils/node_utils.dart';
 import '../../../lint_utils.dart';
+import '../../../metrics/metrics_list/source_lines_of_code/source_code_visitor.dart';
 import '../../../models/internal_resolved_unit_result.dart';
 import '../../../models/issue.dart';
 import '../../../models/severity.dart';
@@ -22,9 +23,11 @@ class PreferExtractingCallbacksRule extends FlutterRule {
       'Prefer extracting the callback to a separate widget method.';
 
   final Iterable<String> _ignoredArguments;
+  final int? _allowedLineCount;
 
   PreferExtractingCallbacksRule([Map<String, Object> config = const {}])
       : _ignoredArguments = _ConfigParser.parseIgnoredArguments(config),
+        _allowedLineCount = _ConfigParser.parseAllowedLineCount(config),
         super(
           id: ruleId,
           severity: readSeverity(config, Severity.style),
@@ -33,7 +36,7 @@ class PreferExtractingCallbacksRule extends FlutterRule {
 
   @override
   Iterable<Issue> check(InternalResolvedUnitResult source) {
-    final visitor = _Visitor(_ignoredArguments);
+    final visitor = _Visitor(source, _ignoredArguments, _allowedLineCount);
 
     source.unit.visitChildren(visitor);