+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_extracting_callbacks/visitor.dart
@@ -3,11 +3,13 @@ part of 'prefer_extracting_callbacks_rule.dart';
 class _Visitor extends SimpleAstVisitor<void> {
   final _expressions = <Expression>[];
 
+  final InternalResolvedUnitResult _source;
   final Iterable<String> _ignoredArguments;
+  final int? _allowedLineCount;
 
   Iterable<Expression> get expressions => _expressions;
 
-  _Visitor(this._ignoredArguments);
+  _Visitor(this._source, this._ignoredArguments, this._allowedLineCount);
 
   @override
   void visitClassDeclaration(ClassDeclaration node) {
@@ -17,7 +19,8 @@ class _Visitor extends SimpleAstVisitor<void> {
       return;
     }
 
-    final visitor = _InstanceCreationVisitor(_ignoredArguments);
+    final visitor =
+        _InstanceCreationVisitor(_source, _ignoredArguments, _allowedLineCount);
     node.visitChildren(visitor);
 
     _expressions.addAll(visitor.expressions);
@@ -27,11 +30,14 @@ class _Visitor extends SimpleAstVisitor<void> {
 class _InstanceCreationVisitor extends RecursiveAstVisitor<void> {
   final _expressions = <Expression>[];
 
+  final InternalResolvedUnitResult _source;
   final Iterable<String> _ignoredArguments;
+  final int? _allowedLineCount;
 
   Iterable<Expression> get expressions => _expressions;
 
-  _InstanceCreationVisitor(this._ignoredArguments);
+  _InstanceCreationVisitor(
+      this._source, this._ignoredArguments, this._allowedLineCount);
 
   @override
   void visitInstanceCreationExpression(InstanceCreationExpression node) {
@@ -44,7 +50,8 @@ class _InstanceCreationVisitor extends RecursiveAstVisitor<void> {
       if (_isNotIgnored(argument) &&
           expression is FunctionExpression &&
           _hasNotEmptyBlockBody(expression) &&
-          !_isFlutterBuilder(expression)) {
+          !_isFlutterBuilder(expression) &&
+          _isLongEnough(expression)) {
         _expressions.add(argument);
       }
     }
@@ -74,4 +81,16 @@ class _InstanceCreationVisitor extends RecursiveAstVisitor<void> {
   bool _isNotIgnored(Expression argument) =>
       argument is! NamedExpression ||
       !_ignoredArguments.contains(argument.name.label.name);
+
+  bool _isLongEnough(Expression expression) {
+    final allowedLineCount = _allowedLineCount;
+    if (allowedLineCount == null) {
+      return true;
+    }
+
+    final visitor = SourceCodeVisitor(_source.lineInfo);
+    expression.visitChildren(visitor);
+
+    return visitor.linesWithCode.length > allowedLineCount;
+  }
 }