+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * chore: restrict `analyzer` version to `>=2.4.0 <4.2.0`.
 * fix: normalize file paths after extraction from analyzed folder.
 * fix: improve context root included files calculation.
+* feat: add [`avoid-banned-imports`](https://dartcodemetrics.dev/docs/rules/common/avoid-banned-imports) rule
 * fix: resolve package with imported analysis options.
 
 ## 4.15.2